Update README.md
I've edited the first part of the documentation here to improve clarity and readability, including making it easier to understand for less technically-inclined users who wish to test the new Corescripts via the release builds.

Please understand that I don't mean to insult you with these changes. This is very well written for someone with English as a second language, and I would like to offer my assistance in helping to polish this and future documentation as well. :)

If you like the idea, I will return to finish refining the remainder of the document at a later time.